fix: linkEntity model tests and other broken structural queries
