+++ packages/hash/api/src/model/knowledge/user.model.ts
@@ -446,6 +446,12 @@ export default class extends EntityModel {
               },
             ],
           },
+          {
+            equal: [{ path: ["version"] }, { parameter: "latest" }],
+          },
+          {
+            equal: [{ path: ["archived"] }, { parameter: false }],
+          },
         ],
       });
